Move doRequest() into its own file in utils/
This makes it easier to use on its own, as suggested in
https://github.com/OctoLinker/browser-extension/issues/280#issuecomment-279037295
